cleaner build
given, ideally, the main python script would be ran through this run.sh
script, it's only fair to assume that if the envionment is set correctly
then pip packages are also installed within it successfully.

this update installs pip packages right after creating virtual
environment, and doesn't check for it for as long as the envionment
already exists.